requirements.txt
numpy>=1.24.0,<1.25.0
tensorflow-cpu>=2.15.0  # Using CPU version as it has better compatibility
librosa>=0.10.1
pyannote.audio>=3.1.1
soundfile>=0.12.1
torch>=2.1.0
noisereduce>=3.0.0
SpeechRecognition>=3.10.0
tqdm>=4.65.0
